big updated added readme file , swagger and .env.example in backend

+++ mobile/lib/core/services/address_injection.dart
@@ -1,5 +1,4 @@
-// Inside initDependencies():
-
+// lib/core/services/address_injection.dart
 import 'package:get_it/get_it.dart';
 import 'package:mobile/features/product/data/datasources/address_remote_datasource.dart';
 import 'package:mobile/features/product/data/repositories/address_repository_impl.dart';
@@ -11,22 +10,43 @@ import 'package:mobile/features/product/domain/usecases/set_default_address.dart
 import 'package:mobile/features/product/presentation/blocs/address_bloc.dart';
 
 void addressRegisterDependencies(GetIt sl) {
-  sl.registerLazySingleton<AddressRemoteDataSource>(
-    () => AddressRemoteDataSourceImpl(client: sl()),
-  );
-  sl.registerLazySingleton<AddressRepository>(
-    () => AddressRepositoryImpl(remoteDataSource: sl(), storageService: sl()),
-  );
-  sl.registerLazySingleton(() => GetAddresses(sl()));
-  sl.registerLazySingleton(() => AddAddress(sl()));
-  sl.registerLazySingleton(() => SetDefaultAddress(sl()));
-  sl.registerLazySingleton(() => DeleteAddress(sl()));
-  sl.registerFactory(
-    () => AddressBloc(
-      getAddresses: sl(),
-      addAddress: sl(),
-      setDefaultAddress: sl(),
-      deleteAddress: sl(),
-    ),
-  );
+  // Data Sources
+  if (!sl.isRegistered<AddressRemoteDataSource>()) {
+    sl.registerLazySingleton<AddressRemoteDataSource>(
+      () => AddressRemoteDataSourceImpl(client: sl()),
+    );
+  }
+
+  // Repositories
+  if (!sl.isRegistered<AddressRepository>()) {
+    sl.registerLazySingleton<AddressRepository>(
+      () => AddressRepositoryImpl(remoteDataSource: sl(), storageService: sl()),
+    );
+  }
+
+  // Use Cases
+  if (!sl.isRegistered<GetAddresses>()) {
+    sl.registerLazySingleton(() => GetAddresses(sl()));
+  }
+  if (!sl.isRegistered<AddAddress>()) {
+    sl.registerLazySingleton(() => AddAddress(sl()));
+  }
+  if (!sl.isRegistered<SetDefaultAddress>()) {
+    sl.registerLazySingleton(() => SetDefaultAddress(sl()));
+  }
+  if (!sl.isRegistered<DeleteAddress>()) {
+    sl.registerLazySingleton(() => DeleteAddress(sl()));
+  }
+
+  // BLoCs
+  if (!sl.isRegistered<AddressBloc>()) {
+    sl.registerFactory(
+      () => AddressBloc(
+        getAddresses: sl(),
+        addAddress: sl(),
+        setDefaultAddress: sl(),
+        deleteAddress: sl(),
+      ),
+    );
+  }
 }
